feat(NewPokemon,NewPokemonUiModel): id 를 Long -> String

+++ android/data/src/main/java/poke/rogue/helper/data/model/NewPokemon.kt
@@ -2,7 +2,7 @@ package poke.rogue.helper.data.model
 
 data class NewPokemon(
     val id: String,
-    val dexNumber: Int,
+    val dexNumber: Long,
     val name: String,
     val imageUrl: String,
     val types: List<Type>,
